+++ packages/js-dpp/lib/test/fixtures/getDataContractFixture.js
@@ -56,7 +56,6 @@ module.exports = function getDataContractFixture(ownerId = randomOwnerId) {
         {
           name: 'index3',
           properties: [
-            { $id: 'asc' },
             { lastName: 'asc' },
           ],
         },
@@ -103,6 +102,7 @@ module.exports = function getDataContractFixture(ownerId = randomOwnerId) {
           prefixItems: [
             {
               type: 'string',
+              maxLength: 100,
             },
           ],
           minItems: 1,
@@ -208,7 +208,6 @@ module.exports = function getDataContractFixture(ownerId = randomOwnerId) {
         {
           name: 'index2',
           properties: [
-            { $id: 'asc' },
             { $ownerId: 'asc' },
             { firstName: 'asc' },
             { lastName: 'asc' },