+++ packages/js-drive/lib/abci/handlers/query/identityIdsByPublicKeyHashesQueryHandlerFactory.js
@@ -6,6 +6,8 @@ const {
   },
 } = require('@dashevo/abci/types');
 
+const cbor = require('cbor');
+
 const {
   v0: {
     GetIdentityIdsByPublicKeyHashesResponse,
@@ -53,8 +55,7 @@ function identityIdsByPublicKeyHashesQueryHandlerFactory(
     if (!blockExecutionContextStack.getLast()) {
       const response = new GetIdentityIdsByPublicKeyHashesResponse();
 
-      response.setIdentityIdsList(publicKeyHashes.map(() => Buffer.alloc(0)));
-
+      response.setIdentityIdsList(publicKeyHashes.map(() => cbor.encode([])));
       response.setMetadata(new ResponseMetadata());
 
       return new ResponseQuery({
@@ -66,7 +67,7 @@ function identityIdsByPublicKeyHashesQueryHandlerFactory(
 
     const identityIds = await Promise.all(
       publicKeyHashes.map(async (publicKeyHash) => (
-        previousPublicKeyToIdentityIdRepository.fetch(publicKeyHash)
+        previousPublicKeyToIdentityIdRepository.fetchBuffer(publicKeyHash)
       )),
     );
 
@@ -87,15 +88,15 @@ function identityIdsByPublicKeyHashesQueryHandlerFactory(
       proof.setRootTreeProof(rootTreeProof);
       proof.setStoreTreeProofs(storeTreeProofs);
     } else {
-      const identityIdBuffers = identityIds.map((identityId) => {
-        if (!identityId) {
-          return Buffer.alloc(0);
+      const idsList = identityIds.map((ids) => {
+        if (!ids) {
+          return cbor.encode([]);
         }
 
-        return identityId.toBuffer();
+        return ids;
       });
 
-      response.setIdentityIdsList(identityIdBuffers);
+      response.setIdentityIdsList(idsList);
     }
 
     return new ResponseQuery({