+++ scripts/configure_dashmate.sh
@@ -4,16 +4,16 @@ set -e
 
 CONFIG_NAME="local"
 
-full_path=$(realpath "$0")
-dir_path=$(dirname "$full_path")
-root_path=$(dirname "$dir_path")
-packages_path="$root_path/packages"
+FULL_PATH=$(realpath "$0")
+DIR_PATH=$(dirname "$FULL_PATH")
+ROOT_PATH=$(dirname "$DIR_PATH")
+PACKAGES_PATH="$ROOT_PATH/packages"
 
-DAPI_REPO_PATH="${packages_path}/dapi"
-DRIVE_REPO_PATH="${packages_path}/js-drive"
+DAPI_REPO_PATH="${PACKAGES_PATH}/dapi"
+DRIVE_REPO_PATH="${PACKAGES_PATH}/js-drive"
 
-yarn dashmate config:set --config=${CONFIG_NAME} platform.dapi.api.docker.build.path "$DAPI_REPO_PATH"
-yarn dashmate config:set --config=${CONFIG_NAME} platform.drive.abci.docker.build.path "$DRIVE_REPO_PATH"
+# build Drive and DAPI from sources
+yarn dashmate config set --config=${CONFIG_NAME} platform.sourcePath "$ROOT_PATH"
 
 # create tenderdash blocks every 10s to speed up test suite
-yarn dashmate config:set --config=${CONFIG_NAME} platform.drive.tenderdash.consensus.createEmptyBlocksInterval "10s"
+yarn dashmate config set --config=${CONFIG_NAME} platform.drive.tenderdash.consensus.createEmptyBlocksInterval "10s"